# [GEOS-11089] Performance penalty adding namespaces while loading catalog