Port GEOS-11089 fix from upstream to gscloud's CatalogPlugin's CatalogInfoLookup.

Use a specialization of `CatalogInfoLookup` for `NamespaceInfo` that speeds up lookup of namespaces by URI.

With this patch and GEOS-11088, starting up from a catalog with 120k workspaces and 600k layers, goes from 18 to 4 minutes with the default geoserver loader, and down to less than a minute with the datadir-catalog-loader community module.

```
                                      before   after
default loader                        1113s    294s
datadir-catalog-loader (4 trheads)     587s     98s
datadir-catalog-loader (8 trheads)     334s     69s
datadir-catalog-loader (16 trheads)    245s     57s
```